backend/requirements.txt
mcp
google-cloud-aiplatform
google-cloud-firestore
google-cloud-pubsub
uvicorn
fastapi
python-dotenv
requests
googlesearch-python
google-adk
google-genai
